unix, windows, darwin : Implemented uv channels
A channel provides a mechanism for two concurrently executing functions
to communicate by passing any value. This value will be passed to all
the receivers that the channel has.

Channel internally holds a mutex and synchronizes the messages it gets.
Because of this, user don't have to acquire a lock when sending or
receiving from channels.